Make script executable and refactor
Add execute permissions to file. Probably why the first attempt failed.
and revert to using file's `-delete` action to remove the files.
Have the `-name` test come before the `-type` test. Why? From the
examples in the man page for find:
"The -name test comes before the -type test in order to
avoid having to call stat(2) on every file."