Convert a number from base 2 to base 10
From the bash manual:

https://www.gnu.org/software/bash/manual/html_node/Shell-Arithmetic.html

Integer constants follow the C language definition, without suffixes or
character constants. Constants with a leading 0 are interpreted as octal
numbers. A leading ‘0x’ or ‘0X’ denotes hexadecimal.

Otherwise, numbers take the form `[base#]n`, where the optional `base`
is a decimal number between 2 and 64 representing the arithmetic base,
and n is a number in that base. If `base#` is omitted, then base 10 is
used